+++ R/calc_precipitation_chirps.R
@@ -79,7 +79,8 @@ calc_precipitation_chirps <- function(years = 1981:2020,
 
     layer_years <- as.numeric(substr(src_names, 13, 17))
     target_chirps <- chirps[[which(layer_years %in% years)]]
-    datetime <- as.Date(paste0(substr(names(target_chirps), 13, 19), ".01"), "%Y.%m.%d")
+    dates <- as.Date(paste0(substr(names(target_chirps), 13, 19), ".01"), "%Y.%m.%d")
+    datetime <- as.POSIXct(paste0(dates, "T00:00:00Z"))
 
     # extract zonal statistics
     results <- select_engine(