+++ R/calc_traveltime.R
@@ -82,7 +82,7 @@ calc_traveltime <- function(engine = "extract", stats = "mean") {
       dplyr::mutate(distances = distances) %>%
       tidyr::pivot_longer(-distances, names_to = "variable") %>%
       dplyr::mutate(
-        datetime = as.Date("2015-01-01"),
+        datetime = as.POSIXct("2015-01-01T00:00:00Z"),
         variable = paste0(distances, "_", variable),
         unit = "minutes"
       ) %>%