+++ R/calc_treecoverloss_emissions.R
@@ -118,7 +118,7 @@ calc_treecoverloss_emissions <- function(years = 2000:2023,
 
     gfw_stats %>%
       dplyr::mutate(
-        datetime = as.Date(paste0(years, "-01-01")),
+        datetime = as.POSIXct(paste0(years, "-01-01T00:00:00Z")),
         variable = "emissions",
         unit = "Mg",
         value = emissions